requirements.txt
streamlit==1.39.0
anthropic==0.39.0
pillow==10.4.0